Implement workspaces
Each workspace takes its own tiling mode, its own nmaster and its own mfact.
These are attached to an existing MetaWorkspace instance via g_object_set_data.
In order to pass g(u)ints into the gpointer that g_object_set_data takes, we
stuff these g(u)ints into gpointers with G(U)INT_TO_POINTER. They are retrieved
with GPOINTER_TO_(U)INT.

No such function was present for a float, so Arek Window Manager now provides
its own GFLOAT_TO_POINTER and GPOINTER_TO_FLOAT. This is the reason for
switching to floats instead of doubles for mfact and mfact_step: it has to work
on both 32bit and 64bit platforms. These functions do not work when the size of
a float and/or a pointer is not 32bits (the latter should not be a problem, as
GLib does not support these platforms).